Lazy-load PDF and ZIP libraries in the client
These libraries are quite large, so we should only load them
when we need them.
