+++ client/src/chunks/export/ExportChunk.ts
@@ -1,0 +1,15 @@
+import { Deck } from "../../state/DeckBuilderModule";
+import { ChunkLoader } from '../ChunkLoader';
+
+export interface ExportChunk {
+  deckToXml(deck: Deck): string;
+  deckToBinderXml(deck: Deck): string;
+  decksToBinderZip(decks: Deck[], names: string[]): Promise<string>;
+  deckToPdf(deck: Deck): void;
+}
+
+export const exportLoader = new ChunkLoader(
+    () =>
+        import(/* webpackChunkName: "export" */ './ExportChunkInternal')
+            .then(chunk => chunk.default)
+);
