+++ client/src/chunks/export/ExportChunkInternal.ts
@@ -1,0 +1,126 @@
+import JSZip from 'jszip';
+import jsPDF from 'jspdf';
+
+import { Deck, BASICS } from '../../state/DeckBuilderModule';
+import { MtgCard } from '../../draft/DraftState';
+import { ExportChunk } from './ExportChunk';
+
+const EXPORT_CHUNK_INTERNAL: ExportChunk = {
+
+  deckToXml(deck: Deck) {
+    let exportStr = XML_HEADER;
+    let mainMap = new Map<number, DeckEntry>();
+    let sideMap = new Map<number, DeckEntry>();
+    for (const card of deck.maindeck.flat()) {
+      if (!card.definition.mtgo) {
+        continue;
+      }
+      incrementQuantity(mainMap, card.definition);
+    }
+    for (const card of deck.sideboard.flat()) {
+      if (!card.definition.mtgo) {
+        continue;
+      }
+      incrementQuantity(sideMap, card.definition);
+    }
+    for (const [mtgo, card] of mainMap) {
+      exportStr += `<Cards CatID=\"${mtgo}\" Quantity=\"${card.quantity}\"`
+          + ` Sideboard=\"false\" Name=\"${card.name}\" />\n`;
+    }
+    for (const [mtgo, card] of sideMap) {
+      exportStr += `<Cards CatID=\"${mtgo}\" Quantity=\"${card.quantity}\"`
+          + ` Sideboard=\"true\" Name=\"${card.name}\" />\n`;
+    }
+    exportStr += "</Deck>";
+    return `data:text/xml;charset=utf-8,${encodeURIComponent(exportStr)}`;
+  },
+
+  deckToBinderXml(deck: Deck): string {
+    return `data:text/xml;charset=utf-8,${encodeURIComponent(deckToBinderXmlContents(deck))}`;
+  },
+
+  async decksToBinderZip(decks: Deck[], names: string[]): Promise<string> {
+    const zip = new JSZip();
+    decks.map(deckToBinderXmlContents).forEach((deckXml, i) => {
+      zip.file(`${names[i]} (${decks[i].maindeck.flat().length + decks[i].sideboard.flat().length}).dek`,
+          deckXml);
+    });
+    return zip.generateAsync({type: "base64"})
+        .then(base64 => `data:application/zip;base64,${base64}`);
+  },
+
+  deckToPdf(deck: Deck) {
+    const pdf = new jsPDF('p', 'in', 'letter');
+    let cardsOnLine = 0;
+    let linesOnPage = 0;
+    const cards = deck.maindeck.flat().concat(deck.sideboard.flat());
+    Promise.all(cards.map(card =>
+        fetch(card.definition.image_uris[0])
+            .then(resp => resp.arrayBuffer())))
+        .then(images => {
+          cards.forEach((card, i) => {
+            pdf.addImage(new Uint8Array(images[i]), 'JPEG',
+                0.25 + cardsOnLine * 2.4,
+                0.25 + linesOnPage * 3.35,
+                2.4, 3.348);
+            cardsOnLine++;
+            if (cardsOnLine === 3) {
+              cardsOnLine = 0;
+              linesOnPage++;
+              if (linesOnPage == 3) {
+                linesOnPage = 0;
+                if (i < cards.length - 1) {
+                  pdf.addPage();
+                }
+              }
+            }
+          });
+          pdf.save('r38export.pdf');
+        });
+  }
+};
+export default EXPORT_CHUNK_INTERNAL;
+
+function deckToBinderXmlContents(deck: Deck) {
+  let exportStr = XML_HEADER;
+  let map = new Map<number, DeckEntry>();
+  for (const card of deck.maindeck.flat()) {
+    if (!card.definition.mtgo || BASICS.includes(card.definition.mtgo)) {
+      continue;
+    }
+    incrementQuantity(map, card.definition);
+  }
+  for (const card of deck.sideboard.flat()) {
+    if (!card.definition.mtgo) {
+      continue;
+    }
+    incrementQuantity(map, card.definition);
+  }
+  for (const [mtgo, card] of map) {
+    exportStr += `<Cards CatID=\"${mtgo}\" Quantity=\"${card.quantity}\"`
+        + ` Sideboard=\"false\" Name=\"${card.name}\" />\n`;
+  }
+  exportStr += "</Deck>";
+  return exportStr;
+}
+
+function incrementQuantity(map: Map<number, DeckEntry>, card: MtgCard) {
+  let entry = map.get(card.mtgo);
+  if (entry == undefined) {
+    entry = {name: card.name, quantity: 0};
+    map.set(card.mtgo, entry);
+  }
+  entry.quantity++;
+}
+
+interface DeckEntry {
+  name: string;
+  quantity: number;
+}
+
+const XML_HEADER =
+    `<?xml version="1.0" encoding="utf-8"?>
+<Deck xmlns:xsd="http://www.w3.org/2001/XMLSchema" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance">
+<NetDeckID>0</NetDeckID>
+<PreconstructedDeckID>0</PreconstructedDeckID>
+`;
