+++ client/src/ui/deckbuilder/DeckBuilderExportMenu.vue
@@ -1,37 +1,39 @@
 <template>
   <div class="_deck-builder-export-menu">
-    <a
-        :href="exportedDecksZip"
-        download="r38export.zip"
-        class="exportButton"
-        v-if="admin && deck"
-        >
-      Export all
-    </a>
-    <a
-        :href="exportedDeck"
-        download="r38export.dek"
-        class="exportButton"
-        v-if="deck"
-        >
-      Export deck
-    </a>
-    <a
-        :href="exportedBinder"
-        download="r38export.dek"
-        class="exportButton"
-        v-if="deck"
-        >
-      Export binder
-    </a>
-    <a
-        @click="exportToPdf"
-        download="r38export.pdf"
-        class="exportButton"
-        v-if="deck"
-        >
-      Print deck
-    </a>
+    <template v-if="libLoaded && deck != null">
+      <a
+          :href="exportedDecksZip"
+          download="r38export.zip"
+          class="exportButton"
+          v-if="admin"
+          >
+        Export all
+      </a>
+      <a
+          :href="exportedDeck"
+          download="r38export.dek"
+          class="exportButton"
+          >
+        Export deck
+      </a>
+      <a
+          :href="exportedBinder"
+          download="r38export.dek"
+          class="exportButton"
+          >
+        Export binder
+      </a>
+      <a
+          @click="exportToPdf"
+          download="r38export.pdf"
+          class="exportButton"
+          >
+        Print deck
+      </a>
+    </template>
+    <div v-else class="loading-message">
+      Loading...
+    </div>
   </div>
 </template>
 
@@ -39,11 +41,21 @@
 import Vue from 'vue';
 import { authStore } from '../../state/AuthStore';
 import { Deck, deckBuilderStore as store } from '../../state/DeckBuilderModule';
-import { decksToBinderZip, deckToBinderXml, deckToPdf, deckToXml } from '../../draft/deckExport';
+import { ExportChunk, exportLoader } from '../../chunks/export/ExportChunk';
 
 export default Vue.extend({
   props: {
     deckIndex: {type: Number},
+  },
+
+  created() {
+    this.loadExportLib();
+  },
+
+  data() {
+    return {
+      libLoaded: false,
+    };
   },
 
   computed: {
@@ -56,35 +68,48 @@ export default Vue.extend({
     },
 
     exportedDeck(): string {
-      return this.deck ? deckToXml(this.deck) : '';
+      return this.deck ? exportLoader.chunk.deckToXml(this.deck) : '';
     },
 
     exportedBinder(): string {
-      return this.deck ? deckToBinderXml(this.deck) : '';
+      return this.deck ? exportLoader.chunk.deckToBinderXml(this.deck) : '';
     },
   },
 
   asyncComputed: {
     async exportedDecksZip(): Promise<string> {
-      return await decksToBinderZip(store.decks, store.names);
+      if (this.libLoaded) {
+        return await exportLoader.chunk.decksToBinderZip(
+            store.decks, store.names);
+      } else {
+        return '';
+      }
     }
   },
 
   methods: {
+    loadExportLib() {
+      exportLoader.load().then(() => {
+        this.libLoaded = true;
+      })
+    },
+
     exportToPdf() {
       if (this.deck) {
-        deckToPdf(this.deck);
+        exportLoader.chunk.deckToPdf(this.deck);
       }
-    }
+    },
   },
 });
 </script>
 
 <style scoped>
 ._deck-builder-export-menu {
-  background: white;
-  box-shadow: #a5a5a5 3px 2px 2px;
-  border: 1px solid #eee;
+  background-color: #FFF;
+  border-radius: 5px;
+  box-shadow: 0px 1px 4px rgba(0, 0, 0, 0.3);
+  border: 1px solid #ccc;
+  min-width: 150px;
 }
 
 .exportButton {
@@ -98,6 +123,11 @@ export default Vue.extend({
 
 .exportButton:hover {
   background: #ddd;
+}
+
+.loading-message {
+  padding: 10px;
+  height: 40px;
 }
 
 </style>
