+++ package.json
@@ -7,7 +7,7 @@
     "start": "cd client_toolchain && npx tsc -p dev_server/ && node dev_server/dist/client_toolchain/dev_server/src/server.js",
     "dev-deps": "(cd client && npm ci); (cd client_toolchain && npm ci);",
     "prod-deps": "(cd client && npm ci --production); (cd client_toolchain && npm ci --production);",
-    "build": "cd client_toolchain && npx webpack --config config/webpack.prod.js --progress --hide-modules",
+    "build": "cd client_toolchain && npx webpack --config config/webpack.prod.js --progress",
     "build-stats": "cd client_toolchain && npx webpack --config config/webpack.prod.js --progress --profile --json > stats.json",
     "deploy": "npm run prod-deps && npm run build",
     "test": "echo \"Error: no test specified\" && exit 1"
